Fix a bug in the building of the dependency graph
In the old code when constructing the dependency graph, which is stored
in the lock file, the package URL and download method were get from the
Nimble packages list. This is not correct because the package could be
not included in it and even it is present in the packages list, if it
is in development mode, its locked version could be from some fork. Now
the package URL and download method are retrieved from the
`PackageInfo` object. For not installed packages, package URL is being
set there, by getting it from the repository directory using the push
URL corresponding to the remote-tracking branch which the current
branch tracks. If there is no such branch the push URL of the default
remote of the VCS is being used. The download method is also set in the
`PackageInfo` object. It is being determined either by querying the
type of the VCS under which is the package directory in the case of not
installed packages or by an added field in the package metadata in the
case of installed ones.

Related to nim-lang/nimble#127